+++ modules/core/src/storage/mod.rs
@@ -1,8 +1,9 @@
+mod file_system;
 pub(crate) mod models;
 pub(crate) mod repository;
 
-use file_system::{self, path_helper::PathHelper};
 use std::sync::mpsc;
+use storage::file_system::path_helper::PathHelper;
 use storage::models::{LapData, LapHeader};
 use storage::repository::Repository;
 use udp::Packet;
